deploy/requirements.txt
# Core backend
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2

# Agentic AI
crewai==0.152.0
crewai-tools==0.55.0
litellm==1.74.3

# Google Gemini SDK
google-genai==0.7.0

# Data & similarity
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2

# Optional (env)
python-dotenv==1.0.1


